refactor: other breaking changes that were pending a major
