+++ libs/rest/src/bucket/Fetch.ts
