+++ libs/rest/src/bucket/index.ts
